backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.110.1
uvicorn[standard]==0.25.0
starlette==0.37.2
pydantic==2.11.9
pydantic_core==2.33.2
pydantic-settings==2.0.3

# Database
motor==3.3.1
pymongo==4.5.0

# GitHub integration
PyGithub==2.8.1

# HTTP and networking
httpx==0.25.2
requests==2.32.5
urllib3==2.5.0

# Authentication and security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Environment and configuration
python-dotenv==1.1.1

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==8.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==25.9.0
isort==6.0.1
flake8==7.3.0
mypy==1.18.2

# Utilities
click==8.3.0
rich==14.1.0
typer==0.18.0

# Type hints
typing-extensions==4.15.0
annotated-types==0.7.0

# Date and time
python-dateutil==2.9.0.post0
pytz==2025.2

# Cryptography
cryptography==46.0.1
PyNaCl==1.6.0

# Other dependencies
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
packaging==25.0
six==1.17.0
sniffio==1.3.1
anyio==4.10.0
h11==0.16.0
watchfiles==1.1.0
